+++ tests/app/opulence/framework/testing/phpunit/console/mocks/ApplicationTestCase.php
@@ -2,9 +2,9 @@
 /**
  * Copyright (C) 2015 David Young
  *
- * Mocks the HTTP application for use in testing
+ * Mocks the console application for use in testing
  */
-namespace Opulence\Tests\Framework\Tests\HTTP\Mocks;
+namespace Opulence\Tests\Framework\Testing\PHPUnit\Console\Mocks;
 
 use Monolog\Logger;
 use Opulence\Applications\Application;
@@ -14,10 +14,9 @@ use Opulence\Applications\Environments\Environment;
 use Opulence\Applications\Paths;
 use Opulence\Applications\Tasks\Dispatchers\Dispatcher as TaskDispatcher;
 use Opulence\Applications\Tasks\TaskTypes;
-use Opulence\Framework\Bootstrappers\HTTP\Requests\RequestBootstrapper;
-use Opulence\Framework\Bootstrappers\HTTP\Routing\RouterBootstrapper;
-use Opulence\Framework\Bootstrappers\HTTP\Views\ViewFunctionsBootstrapper;
-use Opulence\Framework\Tests\HTTP\ApplicationTestCase as BaseApplicationTestCase;
+use Opulence\Framework\Bootstrappers\Console\Commands\CommandsBootstrapper;
+use Opulence\Framework\Bootstrappers\Console\Composer\ComposerBootstrapper;
+use Opulence\Framework\Testing\PHPUnit\Console\ApplicationTestCase as BaseApplicationTestCase;
 use Opulence\IoC\Container;
 use Opulence\IoC\IContainer;
 use Opulence\Tests\Applications\Mocks\MonologHandler;
@@ -26,18 +25,9 @@ class ApplicationTestCase extends BaseApplicationTestCase
 {
     /** @var array The list of bootstrapper classes to include */
     private static $bootstrappers = [
-        RequestBootstrapper::class,
-        RouterBootstrapper::class,
-        ViewFunctionsBootstrapper::class
+        CommandsBootstrapper::class,
+        ComposerBootstrapper::class,
     ];
-
-    /**
-     * @inheritdoc
-     */
-    protected function getGlobalMiddleware()
-    {
-        return [];
-    }
 
     /**
      * @inheritdoc